requirements.txt
pandas
sqlalchemy
python-dotenv
azure-identity
azure-keyvault
woocommerce
pyodbc